updatehub-device-attributes: Add ipinfo-io hook
The new ipinfo-io hook uses the 'ipinfo.io' service to inquire all
information available by the IP and send it as a device attribute.

Besides the changes need for the new hook packaging, it also has been
added to the runtime class, so it can be enabled on the machines.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>