updatehub-device-attribute: Avoid error when building SDK
The generation of SDK based on an image which uses UpdateHub needs to
add the respective `-dev` packages and it was failing as this recipe
does dynamically split each script to a specific sub package leaving
`PN` empty.

To allow it to properly get installed, we set `ALLOW_EMPTY` to `PN` so
the `-dev` dependency is satisfied.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
(cherry picked from commit 1d80ebb5be5b8e7466e25b197387544726eda801)