classes/updatehub-runtime: Allow use of a generic product uid
Create a known product uid, with zeros, to use with UpdateHub
Community Edition Server, this is useful for tests and development
purpose as we don't need to have a real product to test UpdateHub.
Only show a warning message if UPDATEHUB_PRODUCT_UID is set with
zeros and don't skip recipe.

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>