u-boot-updatehub-script: Add recipe
Recipe for generating bootscript file based in a generic template that
aims to be used in all board with UpdateHub support. This is an effort
to change the most part of UpdateHub supported the board to use
distro_bootcmd and this common bootscript.

All variables are documented in the recipe file.

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>