classes/updatehub-image: Display warnings only when handling tasks
The previous code where showing warnings even when not invoking the
uhuarchive or uhupush tasks. This is not optimal as it emit warnings so
often it loses importance.

With this patch we now emit warnings during the tasks invoke, in a
specific task.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
(cherry picked from commit 8c2ce4ccdf4fd9bed5836a49da7777cd8ec0c76d)